Refactor deposit options and improve modal animations
- Added `DepositOptions.native.tsx` for native platform support.
- Created `DepositOptions.web.tsx` for web-specific implementation.
- Updated `DepositOption/index.tsx` to export the web version.
- Improved animation easing in `AnimatedModal.tsx`.
- Cleaned up imports in `index.js` and `DepositOptionModal.tsx`.
