+++ components/DepositOption/index.tsx
@@ -1,3 +1,4 @@
 export { default as DepositOption } from "./DepositOption";
-export { default as DepositOptions } from "./DepositOptions";
+export { default as DepositOptions } from "./DepositOptions.web";
+
 export { default as DepositOptionModal } from "./DepositOptionModal";
