+++ index.js
@@ -4,5 +4,6 @@ if (Platform.OS !== "web") {
 	import("@thirdweb-dev/react-native-adapter");
 }
 
-import "react-native-reanimated";
 import "expo-router/entry";
+import "react-native-reanimated";
+
